Enabled compiler warnings and fixed a warning

+++ Message.C
@@ -795,7 +795,7 @@ void Facility::initStreams(const DestinationPtr &destination) {
         for (int i=0; i<N_IMPORTANCE; ++i)
             streams_.push_back(new Stream(name_, (Importance)i, destination));
     } else {
-        for (int i=0; i<streams_.size(); ++i)
+        for (size_t i=0; i<streams_.size(); ++i)
             streams_[i]->destination(destination);
     }
 }
